% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_theme}
\alias{change_theme}
\title{Switch beetween dark and light theme}
\usage{
change_theme(input, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{input}{result of input switch button}

\item{session}{session}
}
\description{
Switch beetween dark and light theme
}
