% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fluidpage}
\alias{fluidpage}
\alias{navbarpage}
\title{Add external ressources and theme for fluidPage in shiny}
\usage{
fluidpage(..., id = "dark_mode-page")

navbarpage(..., id = "dark_mode-page")
}
\arguments{
\item{...}{params form fluidPage}

\item{id}{id of fluidpage}
}
\value{
fluidpage
}
\description{
Add external ressources and theme for fluidPage in shiny
}
\section{Functions}{
\itemize{
\item \code{navbarpage}: Nav bar page
}}

