% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{script_js}
\alias{script_js}
\title{Js for switch button}
\usage{
script_js(input, id = "dark_mode-page")
}
\arguments{
\item{input}{name of input}

\item{id}{id of fluidpage or navbarpage}
}
\value{
tag script for switch button
}
\description{
Js for switch button
}
