% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_button_theme}
\alias{switch_button_theme}
\title{Insert a switch button light/dark mode.}
\usage{
switch_button_theme(inputId)
}
\arguments{
\item{inputId}{id of input}
}
\value{
A fluidPage object.
}
\description{
Insert a switch button light/dark mode.
}
