% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_dark_shiny}
\alias{theme_dark_shiny}
\title{Create a dark bootstrap theme shiny apps.}
\usage{
theme_dark_shiny(
  bg = "#000000",
  fg = "#15b7d6",
  primary = "#DE633C",
  secondary = "#15b7d6",
  success = "#fd994e",
  info = "#fd5072",
  warning = "#5cadfd",
  danger = "#c4a5fd",
  base_font = "Noto Sans sans-serif",
  font_scale = 1.1,
  enable_gradients = TRUE,
  enable_shadows = TRUE
)
}
\arguments{
\item{bg}{bg}

\item{fg}{fg}

\item{primary}{color primary}

\item{secondary}{color secondary}

\item{success}{color success}

\item{info}{color info}

\item{warning}{color warning}

\item{danger}{color danger}

\item{base_font}{base font}

\item{font_scale}{font scale}

\item{enable_gradients}{gradients}

\item{enable_shadows}{shadows}
}
\value{
A bs_theme function.
}
\description{
Create a dark bootstrap theme shiny apps.
}
